Move typedalises to one place
