+++ app/src/main/java/com/tari/android/wallet/ffi/FFITariWalletAddress.kt
@@ -32,7 +32,8 @@
  */
 package com.tari.android.wallet.ffi
 
-import com.tari.android.wallet.util.EmojiId
+import com.tari.android.wallet.model.Base58
+import com.tari.android.wallet.model.EmojiId
 
 /**
  * Wrapper for native private key type.
