+++ app/src/main/java/com/tari/android/wallet/infrastructure/backup/BackupUtxos.kt
@@ -1,5 +1,5 @@
 package com.tari.android.wallet.infrastructure.backup
 
-import com.tari.android.wallet.ffi.Base58
+import com.tari.android.wallet.model.Base58
 
 data class BackupUtxos(val utxos: List<String>, val sourceBase58: Base58)
