+++ app/src/main/java/com/tari/android/wallet/model/TypeAlias.kt
@@ -4,3 +4,6 @@ import java.math.BigInteger
 
 typealias TxId = BigInteger
 
+typealias EmojiId = String
+
+typealias Base58 = String
