+++ app/src/main/java/com/tari/android/wallet/ui/fragment/contactBook/data/contacts/ContactDto.kt
@@ -3,7 +3,7 @@ package com.tari.android.wallet.ui.fragment.contactBook.data.contacts
 import android.os.Parcelable
 import com.tari.android.wallet.model.TariWalletAddress
 import com.tari.android.wallet.ui.fragment.contactBook.data.ContactAction
-import com.tari.android.wallet.util.EmojiId
+import com.tari.android.wallet.model.EmojiId
 import kotlinx.parcelize.Parcelize
 import java.util.UUID
